Remove FILTER_FLAG_IPV4 from getIP
Summary:
It was reported that when browsing with an IPv6 address one would get errors like

User notice: Argument val must be string. in
Drupal\hys_loop54\Service\Loop54_Utils::escape() (line 9 of
/*/Loop54_Utils.php)

This is likely because the filter used to filter for only IPv4 addresses, which
meant `$the_ip` would be set to `false` when browsing on an IPv6 address, and this
condition was – in turn – not checked by the library, which only checked if
`$the_ip` was `null` before trying to escape it.

Only the last hunk of the patch would solve the problem in the sense of prevent
the error from occurring, but not satisfactorily; there is no reason to limit IP
addresses to only IPv4. So really, this is two fixes for the price of one!

Test Plan: **Very** Careful visual inspection of code.

Reviewers: JoelKall

Reviewed By: JoelKall

Differential Revision: https://phabricator.loop54.com/D38