Release 1.0.3: Fixes for user ID generation and cookie validity
- Generate a more distinctive user ID
- Make user ID cookie valid on all paths of the domain
- Add code example showing how to sort on multiple criteria

DETAILS:

* Make user ID cookie valid on all paths of the domain
* Generate a more distinctive user ID

Due to oversight when implementing `WebClient` in `RemoteClientInfo`, the cookie
used to save the user ID was only valid under the path on which it was created –
this is contrary to the intention that a user should be identified across the
entire domain.

Squeezed in with the fix mentioned above, the generated user ID (which used to
be a plain UUID4) now has a slightly more distinctive visual appearance (most
notably, it begins with the string "UID") in order to be recogniseable if it
ends up in the wrong contexts. This will hopefully make troubleshooting more
pleasant, at the very least.

* Add code example showing how to sort on multiple criteria

This has been supported all along, but with a clunky interface (because there
has not been a good suggestion on how to do it with a fluent interface while
remaining reasonably well-typed.) We received a lot of support requests asking
how to do it, since it wasn't at all obvious how. Therefore it has been added as
a code example.