vVbueqiz: Check also for empty confirmation (#14)
The build server doesn't supply any input (instant EOF on any read) so the confirmation variable was empty, and this got
interpreted as invalid syntax to the test in if, which then went ahead with the deployment anyway! Not what we want. So
we check whether it's empty first.

To be extra safe, we also invert the conditional.