Implement twig templates and fix problem with Loop54User cookie
Loop54User couldn't be created in WebClient because it causes a problem with already sent header. To implement unique userId per visitor Loop54User cookie has to be created earlier for example in controller or template. I implemented the second option because it should be more versatile solution.

I noticed that twig templates was not correctly implemented that's why I did it too.